Rename the survey GUI package to simple
