+++ deepreefmap/gui/app.py
@@ -26,10 +26,10 @@ from deepreefmap.gui.models.management import ModelManagementMixin
 from deepreefmap.gui.runs.past_runs import PastRunsMixin
 from deepreefmap.gui.runs.results import ResultsMixin
 from deepreefmap.gui.runs.loading import RunLoadingMixin
-from deepreefmap.gui.survey.analysis import SurveyAnalysisMixin
-from deepreefmap.gui.survey.batch import SurveyBatchMixin
-from deepreefmap.gui.survey.mode import UiModeMixin
-from deepreefmap.gui.survey.plan import SurveyPlanMixin
+from deepreefmap.gui.simple.analysis import SimpleAnalysisMixin
+from deepreefmap.gui.simple.batch import SimpleBatchMixin
+from deepreefmap.gui.simple.mode import UiModeMixin
+from deepreefmap.gui.simple.plan import SimplePlanMixin
 from deepreefmap.gui.system.panel import SystemPanelMixin
 from deepreefmap.gui.viewer.controls import ViewerControlsMixin
 from deepreefmap.gui.runs.progress import ProgressBarsMixin
@@ -47,9 +47,9 @@ class DeepReefMapWindow(
     ProgressBarsMixin,
     ResultsMixin,
     RunLoadingMixin,
-    SurveyAnalysisMixin,
-    SurveyBatchMixin,
-    SurveyPlanMixin,
+    SimpleAnalysisMixin,
+    SimpleBatchMixin,
+    SimplePlanMixin,
     SystemPanelMixin,
     UiModeMixin,
     ViewerControlsMixin,
