+++ deepreefmap/gui/simple/__init__.py
@@ -1,0 +1,1 @@
+"""Simple-mode GUI: the plan, batch, and analysis workflow with preset settings."""
