--- deepreefmap/gui/survey/__init__.py
@@ -1,1 +1,0 @@
-"""Survey-mode GUI: the simple workflow of plan, batch, and analysis tabs."""
