+++ tests/gui/test_simple_mode.py
